Log the response when using `test:services:trace`
