+++ services/base.spec.js
@@ -3,6 +3,7 @@
 const { expect } = require('chai')
 const { test, given, forCases } = require('sazerac')
 const sinon = require('sinon')
+const trace = require('./trace')
 
 const {
   NotFound,
@@ -98,7 +99,7 @@ describe('BaseService', function() {
       sandbox.restore()
     })
     beforeEach(function() {
-      sinon.stub(DummyService, 'logTrace')
+      sandbox.stub(trace, 'logTrace')
     })
     it('Invokes the logger as expected', async function() {
       const serviceInstance = new DummyService({}, defaultConfig)
@@ -108,13 +109,13 @@ describe('BaseService', function() {
         },
         { queryParamA: '!' }
       )
-      expect(DummyService.logTrace).to.be.calledWithMatch(
+      expect(trace.logTrace).to.be.calledWithMatch(
         'inbound',
         sinon.match.string,
         'Service class',
         'DummyService'
       )
-      expect(DummyService.logTrace).to.be.calledWith(
+      expect(trace.logTrace).to.be.calledWith(
         'inbound',
         sinon.match.string,
         'Named params',
@@ -122,7 +123,7 @@ describe('BaseService', function() {
           namedParamA: 'bar.bar.bar',
         }
       )
-      expect(DummyService.logTrace).to.be.calledWith(
+      expect(trace.logTrace).to.be.calledWith(
         'inbound',
         sinon.match.string,
         'Query params',
