+++ services/service-tester.js
@@ -1,7 +1,9 @@
 'use strict'
 
 const frisby = require('icedfrisby-nock')(require('icedfrisby'))
+const emojic = require('emojic')
 const config = require('../lib/test-config')
+const trace = require('./trace')
 
 /**
  * Encapsulate a suite of tests. Create new tests using create() and register
@@ -47,6 +49,16 @@ class ServiceTester {
       .baseUri(`http://localhost:${config.port}${this.pathPrefix}`)
       .before(() => {
         this.beforeEach()
+      })
+      // eslint-disable-next-line mocha/prefer-arrow-callback
+      .finally(function() {
+        // `this` is the IcedFrisby instance.
+        trace.logTrace(
+          'outbound',
+          emojic.shield,
+          'Response',
+          JSON.parse(this._response.body)
+        )
       })
 
     this.specs.push(spec)
